backend/agents/visual_director.py
"""Agent 3 (Strands, Haiku 4.5): rewrites `broll_queries` on an already-written
screenplay so the stock-footage picker gets better candidates.

The screenwriter has many jobs (verbatim VO, timestamps, hook, on-screen text,
caption, hashtags, and visuals). Visuals get less attention than they deserve.
This agent does one thing: re-score each scene's b-roll queries with a
cinematographer's eye, optimizing for Pexels-friendly texture + motion
vocabulary and a coherent visual flow across scenes.
"""
import os

from strands import Agent
from strands.models import BedrockModel

from models import Screenplay

DIRECTOR_MODEL = os.environ.get(
    "BEDROCK_DIRECTOR_MODEL", "us.anthropic.claude-haiku-4-5-20251001-v1:0"
)

SYSTEM = """You are a stock-footage cinematographer for short-form spiritual reels.

Your ONE job: rewrite each scene's `broll_queries` (3 ordered alternates) so the
Pexels stock search returns cinematic, on-mood footage. Do NOT change any other
field: keep voiceover, on_screen_text, visual, source_start, source_end, start,
end, title, duration_sec, aspect, scenes ordering, caption, hashtags intact.

Heuristics for great queries:

1. Pexels rewards TEXTURE + MOTION + LIGHT + FRAMING keywords.
   Mix these liberally: "macro close up", "slow motion", "aerial", "time lapse",
   "golden hour", "blue hour", "bokeh", "silhouette", "lens flare",
   "ink in water", "smoke swirl", "mist morning", "soft focus".

2. METAPHOR > literal. The podcast is spiritual/philosophical.
   ✅ "ink drop water slow motion" for ignorance spreading
   ✅ "sunrise mountain timelapse" for awakening
   ✅ "single candle flame dark" for stillness
   ✅ "wind long grass golden hour" for letting go
   ❌ "hindu temple", "person praying", "god in sky", "brahmin priest"
   ❌ "man drinking alcohol" (too literal for 'ignorance covering wisdom')
   ❌ "brain with fire" (cringe)
   ❌ "silhouette mountaintop sunset" (wellness-content cliché)

3. Topic-to-visual mappings you should reach for:
   - ego / attachment / letting go   → flowing water, wind, waves, falling leaves
   - ignorance / delusion / confusion → fog, mist, frosted glass, underwater light
   - clarity / awakening / realisation → sunrise, first light, opening eyes macro,
                                           lens flare, dawn mountains
   - action / duty / free will       → hands working, footsteps, runner slow-mo,
                                           machinery, pottery hands
   - cosmic / universal / atman      → starfield, nebula, night sky, spinning galaxy
   - anxiety / overthinking          → city rush time-lapse, flashing signs,
                                           close-up strained eyes
   - stillness / peace               → still pond, snowfall, steady flame, candle

4. VISUAL FLOW across scenes. Vary framing scene-to-scene: close-up → wide →
   aerial/abstract. Don't return three macro shots in a row unless the VO calls
   for it. A reel with one close-up (hands), one wide (landscape), one abstract
   (flowing water) reads better than three of the same.

5. AVOID:
   - Religious iconography (temples, idols, priests, puja, saffron robes).
   - Indian-specific shorthand unless the VO is literally about India.
   - Clips where a human face would dominate the frame during a contemplative
     line — prefer hands, landscapes, objects, or slow-motion abstracts.
   - Stock wellness clichés (silhouette-on-mountaintop-sunset, praying-hands,
     lotus-flower-on-water, woman-arms-outstretched-field).

Each query: 3-5 specific keywords. Be concrete about texture, motion, light.

Also set `broll_query` to the same text as `broll_queries[0]` for back-compat
with older code paths.

Return the FULL screenplay JSON with the same schema, only b-roll queries
changed.
"""


def direct(screenplay: Screenplay) -> Screenplay:
    """Rewrite broll_queries across the screenplay. Everything else is preserved."""
    agent = Agent(
        model=BedrockModel(model_id=DIRECTOR_MODEL, temperature=0.5),
        system_prompt=SYSTEM,
    )
    prompt = (
        "Rewrite the broll_queries in this screenplay according to your rules.\n"
        "Keep every other field identical.\n\n"
        + screenplay.model_dump_json(indent=2)
    )
    return agent.structured_output(Screenplay, prompt)
